perf: improve options for sharing content
